feat(xRPC): support dynamic upstream with upstream_id (#6919)
